requirements.txt
arrow==1.3.0
asgiref==3.9.1
attrs==23.1.0
binaryornot==0.4.4
build==0.10.0
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
cleo==2.0.1
click==8.2.1
cookiecutter==2.6.0
crashtest==0.4.1
distlib==0.3.7
Django==5.2.4
dulwich==0.21.6
et-xmlfile==1.1.0
filelock==3.12.4
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
Jinja2==3.1.6
jsonschema==4.17.3
keyring==24.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.6
numpy==1.26.1
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.10.0
poetry-core==1.7.0
poetry-plugin-export==1.5.0
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.19.2
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0.2
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==14.0.0
shellingham==1.5.3
six==1.16.0
sqlparse==0.5.3
text-unidecode==1.3
tomlkit==0.12.1
trove-classifiers==2023.9.19
types-python-dateutil==2.9.0.20250516
tzdata==2023.3
urllib3==2.0.5
uv==0.7.13
virtualenv==20.24.5
xattr==0.10.1
xlrd==2.0.1
zipp==3.17.0
